+++ kirc-core/src/main/kotlin/de/cmdjulian/kirc/spec/manifest/LayerReference.kt
@@ -2,4 +2,4 @@ package de.cmdjulian.kirc.spec.manifest
 
 import de.cmdjulian.kirc.image.Digest
 
-data class LayerReference(val mediaType: String, val size: UInt, val digest: Digest)
+data class LayerReference(val mediaType: String, val size: Long, val digest: Digest)
