+++ kirc-core/src/main/kotlin/de/cmdjulian/kirc/spec/manifest/ManifestListEntry.kt
@@ -4,6 +4,6 @@ import de.cmdjulian.kirc.image.Digest
 import de.cmdjulian.kirc.spec.Architecture
 import de.cmdjulian.kirc.spec.OS
 
-data class ManifestListEntry(val mediaType: String, val digest: Digest, val size: UInt, val platform: Platform?) {
+data class ManifestListEntry(val mediaType: String, val digest: Digest, val size: Long, val platform: Platform?) {
     data class Platform(val os: OS, val architecture: Architecture, val features: List<String>)
 }
